verificações de superuser nas funções

+++ category/views.py
@@ -9,18 +9,30 @@ from rest_framework.permissions import IsAuthenticated
 
 class CategoryCreate(APIView):
     permission_classes = [IsAuthenticated]
+
     def post(self, request, *args, **kwargs):
+        # Verificar se o usuário é superusuário
+        if not request.user.is_superuser:
+            return Response(
+                {'detail': 'You do not have permission to perform this action.'},
+                status=status.HTTP_403_FORBIDDEN
+            )
+
         serializer = CategorySerializer(data=request.data)
 
         if serializer.is_valid():
             category = serializer.save()
-            return Response({'message': 'Category successfully registered', 'category': serializer.data}, status=status.HTTP_201_CREATED)
+            return Response(
+                {'message': 'Category successfully registered', 'category': serializer.data},
+                status=status.HTTP_201_CREATED
+            )
+
         return Response({'errors': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
-    
+
     def http_method_not_allowed(self, request, *args, **kwargs):
         return Response({
             'detail': 'Method not allowed'
-        },  status=status.HTTP_405_METHOD_NOT_ALLOWED)
+        }, status=status.HTTP_405_METHOD_NOT_ALLOWED)
         
 class CategoryView(APIView):
     permission_classes = [IsAuthenticated]
@@ -52,6 +64,12 @@ class CategoryViewById(APIView):
 class CategoryUpdate(APIView):
     permission_classes = [IsAuthenticated]
     def put(self, request, pk, *args, **kwargs):
+        if not request.user.is_superuser:
+            return Response(
+                {'detail': 'You do not have permission to perform this action.'},
+                status=status.HTTP_403_FORBIDDEN
+            )
+            
         try:
             category = Category.objects.get(pk=pk)
         except Category.DoesNotExist:
@@ -73,6 +91,13 @@ class CategoryUpdate(APIView):
 class CategoryDelete(APIView):
     permission_classes = [IsAuthenticated]
     def delete(self, request, pk, *args, **kwargs):
+        
+        if not request.user.is_superuser:
+            return Response(
+                {'detail': 'You do not have permission to perform this action.'},
+                status=status.HTTP_403_FORBIDDEN
+            )
+            
         try:
             category = Category.objects.get(pk=pk)
             category.delete()
